For TPS feature request 497
Added new config directive disable_external_commands that should be given a boolean TRUE or FALSE (not a string).

Added if statement in submit_nagios_command to check if external commands were disabled. If they were, it simply prints an API error.

Added a new constant in the constants.inc.php that corresponds to the command disabled error.

git-svn-id: https://devhub.nagios.com/svn/nrdp@51 7ae7bfbb-3f01-4008-b2e3-ed148782e3a1